test(spec): Remove specs use with src
We no longer overtest specs with src/ juste with build/
